Refactor mport.list display code to a method.
This will allow us to
avoid the exec call in mport command eventually.

We also avoid the mport_private.h usage now.

Refactored the quiet flag in mport_init to an enum of mportVerbosity.
It has three values and a helper function mport_verbosity that
translates the old bool flags as needed.

MPORT_VQUIET is equilvant to quiet and MPORT_VVERBOSE is equal to
verbose flag. Default is MPORT_VNORMAL